requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
# openai==0.27.7
redis==4.5.5
hiredis==2.2.3
loguru==0.7.0
# tiktoken==0.4.0